Update main.cpp
First thanks for the tutorials so far! Brand new here so just finding any reason to get into code. Both bytes_read and bytes_written seem to be unnecessary variables as the documentation specifies NULL is usable in this instance. Figured it might help reduce complexity to newcomers like myself.